requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# LangChain and RAG dependencies
langchain==0.0.350
langchain-community>=0.0.2
langchain-openai>=0.0.2
openai>=1.3.7

# LlamaIndex for advanced chunking
llama-index>=0.9.0
llama-index-core>=0.9.0

# LightRAG for graph-based RAG
lightrag-hku>=0.0.1
networkx>=3.0
nano-vectordb>=0.1.0

# Google Gemini
google-generativeai>=0.3.0

# Vector databases
faiss-cpu>=1.7.4
chromadb>=0.4.18

# Document processing
pypdf==4.0.1
python-docx==1.1.0
sentence-transformers==2.7.0

# Embeddings and ML
torch>=2.1.1
transformers>=4.36.2
numpy>=1.24.4
huggingface-hub>=0.20.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2